Move denormalize to utils
